requirements.txt
gunicorn==20.0.4
dj-database-url==0.5.0
whitenoise==5.1.0
psycopg2==2.8.5
django==3.0.8
requests==2.24.0
python-dotenv==0.14.0
progress==1.5
sentry-sdk==0.17.8
